server/requirements.txt
flask
flask_jwt_extended
pyyaml
gunicorn
python-dotenv